refactor: create "plonk/circuit/constraint_system/helpers.rs" file
