+++ halo2_frontend/src/plonk/circuit.rs
@@ -1,8 +1,8 @@
 use crate::{
-    circuit::{layouter::SyncDeps, Layouter, Region, Value},
+    circuit::{layouter::SyncDeps, Layouter, Value},
     plonk::{Assigned, Error},
 };
-use halo2_middleware::circuit::{Any, ChallengeMid};
+use halo2_middleware::circuit::Any;
 use halo2_middleware::ff::Field;
 use halo2_middleware::poly::Rotation;
 
@@ -289,412 +289,4 @@ pub trait Circuit<F: Field> {
     /// the caller will be different depending on the context, and they may or
     /// may not expect to have a witness present.
     fn synthesize(&self, config: Self::Config, layouter: impl Layouter<F>) -> Result<(), Error>;
-}
-
-/// An individual polynomial constraint.
-///
-/// These are returned by the closures passed to `ConstraintSystem::create_gate`.
-#[derive(Debug)]
-pub struct Constraint<F: Field> {
-    name: String,
-    poly: Expression<F>,
-}
-
-impl<F: Field> From<Expression<F>> for Constraint<F> {
-    fn from(poly: Expression<F>) -> Self {
-        Constraint {
-            name: "".to_string(),
-            poly,
-        }
-    }
-}
-
-impl<F: Field, S: AsRef<str>> From<(S, Expression<F>)> for Constraint<F> {
-    fn from((name, poly): (S, Expression<F>)) -> Self {
-        Constraint {
-            name: name.as_ref().to_string(),
-            poly,
-        }
-    }
-}
-
-impl<F: Field> From<Expression<F>> for Vec<Constraint<F>> {
-    fn from(poly: Expression<F>) -> Self {
-        vec![Constraint {
-            name: "".to_string(),
-            poly,
-        }]
-    }
-}
-
-/// A set of polynomial constraints with a common selector.
-///
-/// ```
-/// use halo2_middleware::poly::Rotation;
-/// use halo2curves::pasta::Fp;
-/// # use halo2_frontend::plonk::{Constraints, Expression, ConstraintSystem};
-///
-/// # let mut meta = ConstraintSystem::<Fp>::default();
-/// let a = meta.advice_column();
-/// let b = meta.advice_column();
-/// let c = meta.advice_column();
-/// let s = meta.selector();
-///
-/// meta.create_gate("foo", |meta| {
-///     let next = meta.query_advice(a, Rotation::next());
-///     let a = meta.query_advice(a, Rotation::cur());
-///     let b = meta.query_advice(b, Rotation::cur());
-///     let c = meta.query_advice(c, Rotation::cur());
-///     let s_ternary = meta.query_selector(s);
-///
-///     let one_minus_a = Expression::Constant(Fp::one()) - a.clone();
-///
-///     Constraints::with_selector(
-///         s_ternary,
-///         std::array::IntoIter::new([
-///             ("a is boolean", a.clone() * one_minus_a.clone()),
-///             ("next == a ? b : c", next - (a * b + one_minus_a * c)),
-///         ]),
-///     )
-/// });
-/// ```
-///
-/// Note that the use of `std::array::IntoIter::new` is only necessary if you need to
-/// support Rust 1.51 or 1.52. If your minimum supported Rust version is 1.53 or greater,
-/// you can pass an array directly.
-#[derive(Debug)]
-pub struct Constraints<F: Field, C: Into<Constraint<F>>, Iter: IntoIterator<Item = C>> {
-    selector: Expression<F>,
-    constraints: Iter,
-}
-
-impl<F: Field, C: Into<Constraint<F>>, Iter: IntoIterator<Item = C>> Constraints<F, C, Iter> {
-    /// Constructs a set of constraints that are controlled by the given selector.
-    ///
-    /// Each constraint `c` in `iterator` will be converted into the constraint
-    /// `selector * c`.
-    pub fn with_selector(selector: Expression<F>, constraints: Iter) -> Self {
-        Constraints {
-            selector,
-            constraints,
-        }
-    }
-}
-
-fn apply_selector_to_constraint<F: Field, C: Into<Constraint<F>>>(
-    (selector, c): (Expression<F>, C),
-) -> Constraint<F> {
-    let constraint: Constraint<F> = c.into();
-    Constraint {
-        name: constraint.name,
-        poly: selector * constraint.poly,
-    }
-}
-
-type ApplySelectorToConstraint<F, C> = fn((Expression<F>, C)) -> Constraint<F>;
-type ConstraintsIterator<F, C, I> = std::iter::Map<
-    std::iter::Zip<std::iter::Repeat<Expression<F>>, I>,
-    ApplySelectorToConstraint<F, C>,
->;
-
-impl<F: Field, C: Into<Constraint<F>>, Iter: IntoIterator<Item = C>> IntoIterator
-    for Constraints<F, C, Iter>
-{
-    type Item = Constraint<F>;
-    type IntoIter = ConstraintsIterator<F, C, Iter::IntoIter>;
-
-    fn into_iter(self) -> Self::IntoIter {
-        std::iter::repeat(self.selector)
-            .zip(self.constraints)
-            .map(apply_selector_to_constraint)
-    }
-}
-
-/// Helper struct with the parameters required to convert selector assignments into fixed column
-/// assignments.
-pub struct SelectorsToFixed {
-    compress: bool,
-    num_selectors: usize,
-    max_degree: usize,
-    degrees: Vec<usize>,
-}
-
-impl SelectorsToFixed {
-    /// Convert selector assignments into fixed column assignments based on the parameters in
-    /// `SelectorsToFixed`.
-    pub fn convert<F: Field>(&self, selectors: Vec<Vec<bool>>) -> Vec<Vec<F>> {
-        // The number of provided selector assignments must be the number we
-        // counted for this constraint system.
-        assert_eq!(selectors.len(), self.num_selectors);
-
-        if self.compress {
-            let (polys, _) = compress_selectors::process(
-                selectors
-                    .into_iter()
-                    .zip(self.degrees.iter())
-                    .enumerate()
-                    .map(
-                        |(i, (activations, max_degree))| compress_selectors::SelectorDescription {
-                            selector: i,
-                            activations,
-                            max_degree: *max_degree,
-                        },
-                    )
-                    .collect(),
-                self.max_degree,
-                || Expression::Constant(F::ZERO),
-            );
-            polys
-        } else {
-            selectors
-                .into_iter()
-                .map(|selector| {
-                    selector
-                        .iter()
-                        .map(|b| if *b { F::ONE } else { F::ZERO })
-                        .collect::<Vec<_>>()
-                })
-                .collect()
-        }
-    }
-}
-
-/// Gate
-#[derive(Clone, Debug)]
-pub struct Gate<F: Field> {
-    pub(crate) name: String,
-    pub(crate) constraint_names: Vec<String>,
-    pub(crate) polys: Vec<Expression<F>>,
-    /// We track queried selectors separately from other cells, so that we can use them to
-    /// trigger debug checks on gates.
-    pub(crate) queried_selectors: Vec<Selector>,
-    pub(crate) queried_cells: Vec<VirtualCell>,
-}
-
-impl<F: Field> Gate<F> {
-    /// Returns the gate name.
-    pub fn name(&self) -> &str {
-        self.name.as_str()
-    }
-
-    /// Returns the name of the constraint at index `constraint_index`.
-    pub fn constraint_name(&self, constraint_index: usize) -> &str {
-        self.constraint_names[constraint_index].as_str()
-    }
-
-    /// Returns constraints of this gate
-    pub fn polynomials(&self) -> &[Expression<F>] {
-        &self.polys
-    }
-
-    pub fn queried_selectors(&self) -> &[Selector] {
-        &self.queried_selectors
-    }
-
-    pub fn queried_cells(&self) -> &[VirtualCell] {
-        &self.queried_cells
-    }
-}
-
-/// A "virtual cell" is a PLONK cell that has been queried at a particular relative offset
-/// within a custom gate.
-#[derive(Clone, Debug)]
-pub struct VirtualCell {
-    pub column: Column<Any>,
-    pub rotation: Rotation,
-}
-
-impl<Col: Into<Column<Any>>> From<(Col, Rotation)> for VirtualCell {
-    fn from((column, rotation): (Col, Rotation)) -> Self {
-        VirtualCell {
-            column: column.into(),
-            rotation,
-        }
-    }
-}
-
-/// Exposes the "virtual cells" that can be queried while creating a custom gate or lookup
-/// table.
-#[derive(Debug)]
-pub struct VirtualCells<'a, F: Field> {
-    pub(super) meta: &'a mut ConstraintSystem<F>,
-    pub(super) queried_selectors: Vec<Selector>,
-    pub(super) queried_cells: Vec<VirtualCell>,
-}
-
-impl<'a, F: Field> VirtualCells<'a, F> {
-    fn new(meta: &'a mut ConstraintSystem<F>) -> Self {
-        VirtualCells {
-            meta,
-            queried_selectors: vec![],
-            queried_cells: vec![],
-        }
-    }
-
-    /// Query a selector at the current position.
-    pub fn query_selector(&mut self, selector: Selector) -> Expression<F> {
-        self.queried_selectors.push(selector);
-        Expression::Selector(selector)
-    }
-
-    /// Query a fixed column at a relative position
-    pub fn query_fixed(&mut self, column: Column<Fixed>, at: Rotation) -> Expression<F> {
-        self.queried_cells.push((column, at).into());
-        Expression::Fixed(FixedQuery {
-            index: Some(self.meta.query_fixed_index(column, at)),
-            column_index: column.index,
-            rotation: at,
-        })
-    }
-
-    /// Query an advice column at a relative position
-    pub fn query_advice(&mut self, column: Column<Advice>, at: Rotation) -> Expression<F> {
-        self.queried_cells.push((column, at).into());
-        Expression::Advice(AdviceQuery {
-            index: Some(self.meta.query_advice_index(column, at)),
-            column_index: column.index,
-            rotation: at,
-        })
-    }
-
-    /// Query an instance column at a relative position
-    pub fn query_instance(&mut self, column: Column<Instance>, at: Rotation) -> Expression<F> {
-        self.queried_cells.push((column, at).into());
-        Expression::Instance(InstanceQuery {
-            index: Some(self.meta.query_instance_index(column, at)),
-            column_index: column.index,
-            rotation: at,
-        })
-    }
-
-    /// Query an Any column at a relative position
-    pub fn query_any<C: Into<Column<Any>>>(&mut self, column: C, at: Rotation) -> Expression<F> {
-        let column = column.into();
-        match column.column_type() {
-            Any::Advice => self.query_advice(Column::<Advice>::try_from(column).unwrap(), at),
-            Any::Fixed => self.query_fixed(Column::<Fixed>::try_from(column).unwrap(), at),
-            Any::Instance => self.query_instance(Column::<Instance>::try_from(column).unwrap(), at),
-        }
-    }
-
-    /// Query a challenge
-    pub fn query_challenge(&mut self, challenge: Challenge) -> Expression<F> {
-        Expression::Challenge(challenge)
-    }
-}
-
-/// A selector, representing a fixed boolean value per row of the circuit.
-///
-/// Selectors can be used to conditionally enable (portions of) gates:
-/// ```
-/// use halo2_middleware::poly::Rotation;
-/// # use halo2curves::pasta::Fp;
-/// # use halo2_frontend::plonk::ConstraintSystem;
-///
-/// # let mut meta = ConstraintSystem::<Fp>::default();
-/// let a = meta.advice_column();
-/// let b = meta.advice_column();
-/// let s = meta.selector();
-///
-/// meta.create_gate("foo", |meta| {
-///     let a = meta.query_advice(a, Rotation::prev());
-///     let b = meta.query_advice(b, Rotation::cur());
-///     let s = meta.query_selector(s);
-///
-///     // On rows where the selector is enabled, a is constrained to equal b.
-///     // On rows where the selector is disabled, a and b can take any value.
-///     vec![s * (a - b)]
-/// });
-/// ```
-///
-/// Selectors are disabled on all rows by default, and must be explicitly enabled on each
-/// row when required:
-/// ```
-/// use halo2_frontend::circuit::{Chip, Layouter, Value};
-/// use halo2_frontend::plonk::{Advice, Fixed, Error, Column, Selector};
-/// use halo2_middleware::ff::Field;
-///
-/// struct Config {
-///     a: Column<Advice>,
-///     b: Column<Advice>,
-///     s: Selector,
-/// }
-///
-/// fn circuit_logic<F: Field, C: Chip<F>>(chip: C, mut layouter: impl Layouter<F>) -> Result<(), Error> {
-///     let config = chip.config();
-///     # let config: Config = todo!();
-///     layouter.assign_region(|| "bar", |mut region| {
-///         region.assign_advice(|| "a", config.a, 0, || Value::known(F::ONE))?;
-///         region.assign_advice(|| "a", config.b, 1, || Value::known(F::ONE))?;
-///         config.s.enable(&mut region, 1)
-///     })?;
-///     Ok(())
-/// }
-/// ```
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash)]
-pub struct Selector(pub usize, pub(crate) bool);
-
-impl Selector {
-    /// Enable this selector at the given offset within the given region.
-    pub fn enable<F: Field>(&self, region: &mut Region<F>, offset: usize) -> Result<(), Error> {
-        region.enable_selector(|| "", self, offset)
-    }
-
-    /// Is this selector "simple"? Simple selectors can only be multiplied
-    /// by expressions that contain no other simple selectors.
-    pub fn is_simple(&self) -> bool {
-        self.1
-    }
-
-    /// Returns index of this selector
-    pub fn index(&self) -> usize {
-        self.0
-    }
-
-    /// Return expression from selector
-    pub fn expr<F: Field>(&self) -> Expression<F> {
-        Expression::Selector(*self)
-    }
-}
-
-/// A challenge squeezed from transcript after advice columns at the phase have been committed.
-#[derive(Clone, Copy, Debug, Eq, PartialEq, Hash)]
-pub struct Challenge {
-    pub index: usize,
-    pub(crate) phase: u8,
-}
-
-impl Challenge {
-    /// Index of this challenge.
-    pub fn index(&self) -> usize {
-        self.index
-    }
-
-    /// Phase of this challenge.
-    pub fn phase(&self) -> u8 {
-        self.phase
-    }
-
-    /// Return Expression
-    pub fn expr<F: Field>(&self) -> Expression<F> {
-        Expression::Challenge(*self)
-    }
-}
-
-impl From<Challenge> for ChallengeMid {
-    fn from(val: Challenge) -> Self {
-        ChallengeMid {
-            index: val.index,
-            phase: val.phase,
-        }
-    }
-}
-
-impl From<ChallengeMid> for Challenge {
-    fn from(c: ChallengeMid) -> Self {
-        Self {
-            index: c.index,
-            phase: c.phase,
-        }
-    }
 }
